+++ data_objects/size.hpp
@@ -5,17 +5,17 @@
 class Size
 {
 public:
-    Size(int width, int height);
-    int width() const;
-    void setWidth(int width);
-    int height() const;
-    void setHeight(int height);
+    Size(float width, float height);
+    float width() const;
+    void setWidth(float width);
+    float height() const;
+    void setHeight(float height);
 
     bool isValid() const;
 
 private:
-    int width_;
-    int height_;
+    float width_;
+    float height_;
 };
